Moved cartesian mesh mapper implementation into own cpp file.
