+++ include/neso_particles.hpp
@@ -22,7 +22,6 @@
 #include "neso_particles/containers/resource_stack.hpp"
 #include "neso_particles/containers/rng/rng.hpp"
 #include "neso_particles/containers/sym_vector.hpp"
-#include "neso_particles/containers/sym_vector_impl.hpp"
 #include "neso_particles/containers/tuple.hpp"
 #include "neso_particles/device_functions.hpp"
 #include "neso_particles/domain.hpp"
@@ -51,10 +50,6 @@
 #include "neso_particles/utility.hpp"
 #include "neso_particles/utility_mesh_hierarchy_plotting.hpp"
 // Implementations
-#include "neso_particles/containers/sym_vector_pointer_cache_dispatch_impl.hpp"
-#include "neso_particles/departing_particle_identification_impl.hpp"
-#include "neso_particles/global_mapping_impl.hpp"
-#include "neso_particles/loop/particle_loop_impl.hpp"
-#include "neso_particles/particle_group_impl.hpp"
+#include "neso_particles/common_impl.hpp"
 
 #endif
