+++ include/neso_particles/common_impl.hpp
@@ -1,0 +1,11 @@
+#ifndef _NESO_PARTICLES_COMMON_IMPL_HPP_
+#define _NESO_PARTICLES_COMMON_IMPL_HPP_
+
+#include "containers/sym_vector_impl.hpp"
+#include "containers/sym_vector_pointer_cache_dispatch_impl.hpp"
+#include "departing_particle_identification_impl.hpp"
+#include "global_mapping_impl.hpp"
+#include "loop/particle_loop_impl.hpp"
+#include "particle_group_impl.hpp"
+
+#endif
