+++ src/cell_dat_move.cpp
@@ -1,9 +1,5 @@
 #include <neso_particles/cell_dat_move.hpp>
-#include <neso_particles/departing_particle_identification_impl.hpp>
-#include <neso_particles/global_mapping_impl.hpp>
-#include <neso_particles/loop/particle_loop.hpp>
-#include <neso_particles/loop/particle_loop_impl.hpp>
-#include <neso_particles/particle_group_impl.hpp>
+#include <neso_particles/common_impl.hpp>
 
 namespace NESO::Particles {
