+++ src/particle_io.cpp
@@ -1,7 +1,4 @@
-#include <neso_particles/departing_particle_identification_impl.hpp>
-#include <neso_particles/global_mapping_impl.hpp>
-#include <neso_particles/loop/particle_loop_impl.hpp>
-#include <neso_particles/particle_group_impl.hpp>
+#include <neso_particles/common_impl.hpp>
 #include <neso_particles/particle_io.hpp>
 
 #ifdef NESO_PARTICLES_HDF5
